requirements.txt
pydantic==2.9.2
typer==0.12.5
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pyyaml==6.0.2
sqlite-utils==3.37
fastembed==0.3.5
networkx==3.3
restrictedpython==7.4
numpy==2.1.1
scikit-learn==1.5.2
tqdm==4.66.5
tenacity==9.0.0

